+++ source/define/constant/FUNCTION.ts
@@ -1,0 +1,4 @@
+/**
+  It represents the function expression.
+*/
+export const FUNCTION: string = 'function';