+++ source/define/index.ts
@@ -1,0 +1,8 @@
+/**
+  define
+  ------------------------------------------------------------------------------
+  Central repository for constant values used in general.
+  ------------------------------------------------------------------------------
+*/
+
+export { FUNCTION } from './constant/FUNCTION';